requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
et_xmlfile==2.0.0
fonttools==4.59.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mne==1.10.1
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
statsmodels==0.14.5
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.17.4
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
